AI CAP/requirements.txt
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
docx2pdf==0.1.8
h11==0.16.0
idna==3.10
lxml==6.0.2
outcome==1.3.0.post0
pycparser==2.23
PySocks==1.7.1
python-docx==1.2.0
requests==2.32.5
selenium==4.35.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
fastapi==0.115.4
uvicorn[standard]==0.30.6
PyYAML==6.0.2
python-multipart==0.0.9
jinja2==3.1.4